requirements.txt
pip==23.2.1
pandas==2.1.0
google-api-python-client==2.61.0
numpy==1.25.2
google-auth-httplib2==0.1.0
plotly==5.17.0
streamlit==1.26.0
pillow==9.2.0
ipykernel==6.25.2